feat: add many improvements and create all tests

+++ src/Ambev.DeveloperEvaluation.Application/Carts/DeleteCart/DeleteCartHandler.cs
@@ -1,4 +1,5 @@
 using Ambev.DeveloperEvaluation.Domain.Repositories.Mongo;
+using FluentValidation;
 using MediatR;
 
 namespace Ambev.DeveloperEvaluation.Application.Carts.DeleteCart
@@ -14,6 +15,12 @@ namespace Ambev.DeveloperEvaluation.Application.Carts.DeleteCart
 
         public async Task<bool> Handle(DeleteCartCommand command, CancellationToken cancellationToken)
         {
+            var cart = await _repository.GetByIdAsync(command.Id, cancellationToken);
+            if (cart == null)
+            {
+                throw new ValidationException("Cart not found.");
+            }
+
             return await _repository.DeleteAllCartAsync(command.Id, cancellationToken);
         }
     }
